requirements.txt
asttokens==2.4.1
Babel==2.14.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
colorchooser==1.0
comm==0.2.0
contourpy==1.2.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
debugpy==1.8.0
decorator==5.1.1
et-xmlfile==1.1.0
executing==2.0.1
Flask==3.0.2
fonttools==4.50.0
greenlet==3.0.3
ipykernel==6.26.0
ipython==8.17.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.3
openpyxl==3.1.2
packaging==23.2
pandas==2.2.1
parso==0.8.3
pillow==10.3.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
pure-eval==0.2.2
Pygments==2.16.1
pyparsing==3.1.2
python-dateutil==2.8.2
pytz==2024.1
pywin32==306
pyzmq==25.1.1
six==1.16.0
SQLAlchemy==2.0.25
stack-data==0.6.3
tkcalendar==1.6.1
tornado==6.3.3
traitlets==5.13.0
ttkthemes==3.2.2
typing_extensions==4.9.0
tzdata==2024.1
wcwidth==0.2.9
Werkzeug==3.0.1
